Changed timezone to be float in FacebookProfile

+++ spring-social-facebook/src/main/java/org/springframework/social/facebook/api/impl/json/FacebookProfileMixin.java
@@ -62,7 +62,7 @@ abstract class FacebookProfileMixin {
 	String thirdPartyId;
 	
 	@JsonProperty("timezone")
-	Integer timezone;
+	Float timezone;
 	
 	@JsonProperty("updated_time")
 	Date updatedTime;
