+++ spring-social-facebook/src/test/java/org/springframework/social/facebook/api/UserTemplateTest.java
@@ -45,7 +45,7 @@ public class UserTemplateTest extends AbstractFacebookApiTest {
 		assertEquals("cwalls@vmware.com", profile.getEmail());
 		assertEquals("http://www.facebook.com/habuma", profile.getLink());
 		assertEquals("xyz123abc987", profile.getThirdPartyId());
-		assertEquals(Integer.valueOf(-5), profile.getTimezone());  // should be -6 ???
+		assertEquals(Float.valueOf(-5), profile.getTimezone());  // should be -6 ???
 		assertEquals(toDate("2010-08-22T00:01:59+0000"), profile.getUpdatedTime());
 		assertTrue(profile.isVerified());
 		assertEquals("Just some dude", profile.getAbout());
